TesteComunicacaoUarts/requirements.txt
pyserial
pynmea2